% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CF}
\alias{CF}
\title{Classical F test for mixed effects model}
\usage{
CF(data, a, b, mn)
}
\arguments{
\item{data}{The data in long format (see the example in function  dataformat_wide_to_long( )).}

\item{a}{The number of treatments.}

\item{b}{The number of time points or repeated measurements per subject.}

\item{mn}{The vector of sample sizes in treatments.}
}
\value{
a list of p-values from mixed effect model for the test of no main
   treatment effect (palpha), test of no main time effect (pbeta),
   test of no interaction effect between treatment and time (pgamma),
   and test of no simple effect of treatment (pphi).
}
\description{
#' This function conducts the test of main effect of treatment, interaction effect of treatment and time,  main effect of time,
and simple effect of treament based on mixed effects models
}
