% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cov_squared_jackknife}
\alias{cov_squared_jackknife}
\title{Jackknife estimate of the squared covariance between two time points}
\usage{
cov_squared_jackknife(x, y)
}
\arguments{
\item{x}{a vector of i.i.d. observations}

\item{y}{a second vector of i.i.d. observations}
}
\value{
Jackknife estimate of the squared covariance
}
\description{
This function provides the Jackknife estimate of the squared
covariance using i.i.d. obervations from one variable given in vector \code{x}
and i.i.d. observations from a second variable given in vector \code{y}.
}
