% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fun.sigijj12}
\alias{fun.sigijj12}
\title{Unbiased estimate of  $sigma_{ijj1}^2$}
\usage{
fun.sigijj12(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
A list containing two variables sigmaijj12 and ssijj1 in it.
      The sigmaijj12 variable gives an unbiased estimate of $sigma_{ijj1}^2$.
      The ssijj1 variable gives an unbiased estimate of $sigma_{ijj} sigma_{ij_1j_1}$.
}
\description{
This function calculatess an unbiased estimate of  $sigma_{ijj1}^2$ using the u-statisitic of vectors
$x=(x_1, x_2, ..., x_{ni})$ and  $y=(y_1, y_2, ..., y_{ni})$,
 where $X_j$ and $Y_j$ are correlated, but $X_j$ and $Y_{j1}$ are independent if $j ne j1$.
Note: $sum_{k1 ne k2 ne k3 ne k4} (x_{k1}-x_{k2)})(y_{k1}-y_{k2)}) (x_{k3}-x_{k4)}) (y_{k3}-y_{k4)})$ is an
unbiased est. of  $4*ni*(ni-1)*(ni-2)*(ni-3) [E(X_{ijk}-mu_{ij} u_{ij1k}) ]^2$.
}
