% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sigma4}
\alias{sigma4}
\title{Unbiased estimate of squared variance sigma^4 using C routine}
\usage{
sigma4(y)
}
\arguments{
\item{y}{a numeric vector}
}
\value{
a numeric value
}
\description{
To calculate unbiased estimate of \eqn{\sigma^4} using C routine sigma4c. This
version gives the same return result as sigma4R but is much faster than sigma4R for large n.
}
\examples{
{
\dontrun{
  y = rnorm(10)
  sigma4(y)
 }
}

}
