% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sigma4bootstrap}
\alias{sigma4bootstrap}
\title{Bootstrap estimate of $sigma^4$ using an i.i.d. sample}
\usage{
sigma4bootstrap(x)
}
\arguments{
\item{x}{is a vector of i.i.d. observations.}
}
\value{
Bootstrap estimate of $ sigma^4$, where $sigma^2$ is the variance.
}
\description{
This function takes a random sample and approximates an unbiased estimate of the squared variance based on
Bootstrap estimate.
}
\examples{
  x=stats::rnorm(10)
  sigma4bootstrap(x)
}
